fix template, finish jsDoc
