+++ serverSide/.jsdocrc
@@ -22,6 +22,6 @@
         "encoding": "utf8",
         "private": true,
         "recurse": true,
-        "template": "./node_modules/minami"
+        "template": "./node_modules/tui-jsdoc-template"
     }
 }
