+++ serverSide/package.json
@@ -7,7 +7,8 @@
     "start": "node lib/server.js",
     "test": "node_modules/.bin/mocha test/unit/**/**.spec.js --exit",
     "lint": "eslint --fix --ext .js .",
-    "doc": "jsdoc --configure .jsdocrc"
+    "doc": "npm run generate-doc && opn docs/index.html",
+    "generate-doc": "rimraf docs && jsdoc --configure .jsdocrc"
   },
   "author": "",
   "license": "ISC",
@@ -27,7 +28,9 @@
     "eslint": "^4.18.2",
     "eslint-config-prettier": "^2.9.0",
     "jsdoc": "^3.5.5",
-    "minami": "^1.2.3",
-    "mocha": "^5.0.4"
+    "mocha": "^5.0.4",
+    "opn-cli": "^3.1.0",
+    "rimraf": "^2.6.2",
+    "tui-jsdoc-template": "^1.2.2"
   }
 }
